[MRH] Added forbiddenGroup.txt file to Birad_recombination rxn family folder.
Running a condition.txt file with CO and C2H2 led to an arc PositionOccupiedException.  Take the following chemgraph as an example:

1 O 0 {2,D}
2 C 1 {1,D} {3,S}
3 C 1 {2,S} {4,S} {6,S}
4 C 0 {3,S} {5,D}
5 C 0 {4,D} {6,S}
6 C 1 {5,S} {3,S}

A substructure of this chemgraph is a four-member ring.  Furthermore, adjacent atoms have radicals.  RMG tries to form a bond between atoms 3 and 6 based on the Birad_recombination rxn family (because they are 3 atoms apart).  However, when trying to do so, RMG recognize an arc between the 2 atoms already exists and throws the exception.

I have added 3 forbidden structures to account for this.