Quite a bit of work towards generating reactions for a given species or pair of species.
Added a new class ReactionRecipe used for carrying out reactions on Species objects.

A modified kinetics database is now used in an effort to eliminate reaction family hardcoding. 

Added merge() and split() functions to the ChemGraph class to work with multiple graphs in the same ChemGraph object. For example, this is useful when converting reactants to products; the reactants are combined to a single structure, the reaction recipe is performed, and the resulting products split into individual structures.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@32 37378900-1858-4726-a06b-e54562fdc990