Added species.makeNewSpecies() and reaction.makeNewReaction() functions.
The makeNewSpecies() and makeNewReaction() functions should be called whenever attempting to create a new species or reaction. They check to make sure the species or reaction does not already exist before creating a new one, and return the existing species or reaction if found. Lists of existing species and reactions are maintained in species.speciesList and reaction.reactionList, although I would prefer the user not interact directly with these, but instead through intermediate functions in the species and reaction modules.

Also turned off the OpenBabel warning that is printed when making an InChI when stereochemistry is undefined.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@34 37378900-1858-4726-a06b-e54562fdc990