Some more work towards estimating the kinetics of individual reactions.
Added line to each kinetics library specifying the number of labels used in each line.

Kinetics data is now read and converted to ArrheniusEPKinetics objects; like the thermo database, this slows down the process as unit conversions take place.

Added/renamed functions for working with the labeled atoms of a structure.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@39 37378900-1858-4726-a06b-e54562fdc990