Changed Reaction.getRate() to use k*(C_A * C_B - 1/Keq * C_C * C_D) as the rate expression, i.e. only get the kinetics for the forward reaction.
Added a cache attribute bestKinetics to Reaction for storing the best kinetics for a forward rate. The full getBestKinetics() is now executed the when bestKinetics is None or when the temperature goes outside the valid range of bestKinetics.

Fixed bug in getEnthalpy() and getEntropy() caused by a Quantity object assignment occurring by reference when a copy was needed.

ODE solver is still very slow, as speed gain from use of cache is offset by speed loss from having to evaluate the equilibrium constant every iteration.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@49 37378900-1858-4726-a06b-e54562fdc990