Pruned nodes from the reaction family trees that were unnecessary, and added code to identify unneeded nodes.
A node was deemed unnecessary if it and all of its descendents do not 
contribute any data to the library and it is not the child of a node with a
structure defined by a union.

Only the trees have been pruned at this time; the dictionaries continue to
hold all original nodes. That may be rectified in a future commit.



git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@84 37378900-1858-4726-a06b-e54562fdc990