Converted rmg.chem and rmg.graph modules to use Cython for speed.
In order for RMG to run, you will need to compile the Cython code. This is
best done via running the following command in the source/ folder:

python setup.py build_ext --inplace

Strictly speaking, you do not need Cython installed if you are only building
the chem and graph modules without modifying them, as the *.c files
corresponding to the *.pyx files are also in version control. Successful
compilation will result in *.so files on Linux, or *.pyd files on Windows (the
latter of which are actually DLL files, just with a different extension). You
can then use these modules as if they were pure Python -- only faster!



git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@110 37378900-1858-4726-a06b-e54562fdc990