Rewrote group fitting scheme.
Top nodes are excluded from the group fitting so they end up with values of 0, and a constant has been added.
The method for constructing the matrices has also been changed  - it now creates lists of lists then converts to arrays.
The output is now organised as a tree so you can compare siblings' and parents' values.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@111 37378900-1858-4726-a06b-e54562fdc990