Changed the atom types to reflect the definitions actually used in the RMG database, not my misunderstanding of them. In particular, Cd is now a carbon with one double bond and two single bonds, and there is no Cds.
Fixed a couple of instances of not mutually exclusive functional group
definitions in the thermo radical data. Currently a warning is used to watch
for this aberrant behavior, although it will not cause RMG to fail. Ironically,
this warning is catching a significant number of potential issues in the
kinetics databases; these need looking into.

Also completed the new thermo test for 1,3-hexadiene (C6H10) and the possible 
1,3-hexadienyl (C6H9) radicals. These match the RMG Java values to the desired
level of precision.





git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@131 37378900-1858-4726-a06b-e54562fdc990