[MRH] Updated H_Abstraction tree.
MRH had separate discussions with JWA & RW and with WHG.  The initial problem was that Y_rad_birad did not have exclusive children (Y_rad and Y_2centeradbirad were the culprits).  The temporary fix was done by MRH (removing the Y_2centerradbirad and placing the O2d leaf as a child of Y_rad) but the permanent fix has now been implemented: making Y_2centeradbirad a child of Y_rad.