Moved calls to __VF2_terminals again.
As before, we pass around the terminals
dicts from function to function, but now we use them as the basis for the new
terminals dicts. I.e. there's a new function __VF2_new_terminals which takes the
old terminals and the vertex that's just been added to the map, and returns an
updated copy of the terminals. It saves on having to recalculate a lot of the
existing terminals by copying the old one and updating. Seems to help speed.
(note Test2 is still pretty random)

 Test1 took 0.127 milliseconds ([0.13517498970031738, 0.14175105094909668, 0.14188408851623535, 0.127471923828125, 0.1446070671081543, 0.14035606384277344, 0.13711190223693848, 0.14260506629943848, 0.13648295402526855, 0.13821792602539062])
 Test2 took 0.068 seconds ([0.067527055740356445, 0.25696086883544922, 5.7447900772094727, 2.4298949241638184, 63.96557879447937, 20.726117849349976, 5.329902172088623, 8.3031630516052246, 21.036625862121582, 11.340364933013916])

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@160 37378900-1858-4726-a06b-e54562fdc990