Trying to get more of the graph algorithm code into Cython for speed.
Not actually sure how much faster this is though.

Currently a lot of graph components (vertices and edges) are ctyped as
chem.Atom and chem.Bond. It is inelegant, as the previously general graph class
now only works for atoms and bonds :-/ , but it works!
The graphtest fails (can't convert graph.Vertex to chem.Atom) but everything else works.

I have now just read http://docs.cython.org/docs/pyrex_differences.html#more-friendly-type-casting
which is probably a better way to do it.

Line 6740 of graph.c was a PyObject_GetAttr lookup to find the 'equivalent' method of a vertex,
and was taking ~10% of the cpu time for the whole graph.c file in 10 minutes of a diesel
simulation. Hopefully it now gets the chem.Atom.equivalent c function directly.

Seems a bit faster:

 Test1 took 0.121 milliseconds ([0.12776803970336914, 0.12065291404724121, 0.13396096229553223, 0.20338296890258789, 0.13372921943664551, 0.12780594825744629, 0.14219307899475098, 0.12972497940063477, 0.12891197204589844, 0.14234519004821777])
 Test2 took 0.407 seconds ([0.40694594383239746, 2.2110140323638916, 0.9170689582824707, 1.50571608543396, 15.07253885269165])

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@162 37378900-1858-4726-a06b-e54562fdc990