Replaced a lookup of iteritems() with possibly many lookups of keys in a dict.
Not sure if this will be faster or slower in the real world.
Certainly the old version was spending a lot of time getting and calling the
iteritems method via the Python API, which the new version should avoid
by just iterating over the keys, but the new version will then have to look
up the value for each key.

Preliminary results inconclusive:
 Test1 took 0.116 milliseconds ([0.12294316291809082, 0.11574697494506836, 0.12749099731445312, 0.19361186027526855, 0.1270139217376709, 0.1213231086730957, 0.13773584365844727, 0.12447905540466309, 0.12228989601135254, 0.13708782196044922])
 Test2 took 0.379 seconds ([0.37934994697570801, 2.0756270885467529, 0.85535001754760742, 1.4016768932342529, 14.522458076477051])

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@163 37378900-1858-4726-a06b-e54562fdc990