Sort the list of terminals/atoms before proposing pairs in __VF2_pairs.
Notice that the VF2 algorithm (Cordella et al., 2001) it specifies "any total
ordering criterion could be used" and that this definition "ensures that the
search algorithm never visit the same state twice."

Because keys in a dict are unordered, it is likely that the old algorithm was
visiting the same state more than once, leading to such long search times.

The proposed sorting as implemented here uses a function of the connectivity of
the atoms, so that more highly connected vertices are proposed first (and are
proposed to be matched with highly connected vertices). Strangely, Test1 (small
graph) seems to be faster when sorted the other way around (less connections
first) but Test 2 (large graph) is definitely faster this way.

Here are the times. Notice the units are now milliseconds for both:

 Test1 took 0.268 milliseconds ([0.27110791206359863, 0.27913999557495117, 0.26855087280273438, 0.27149081230163574, 0.26975297927856445, 0.26936221122741699, 0.27486085891723633, 0.2958219051361084, 0.27506780624389648, 0.26831197738647461])
 Test2 took 0.733 milliseconds ([0.74712204933166504, 0.74610590934753418, 0.73255395889282227, 0.76121091842651367, 0.74857187271118164, 0.74539494514465332, 0.75653195381164551, 0.73595499992370605, 0.74848484992980957, 0.80862212181091309])

n.b. Currently the graph.c code spends ~63% of its time doing the sorting,
called via the python api

L. P. Cordella, P. Foggia, C. Sansone, M. Vento
"An improved algorithm for matching large graphs" , Proc. of the 3rd IAPR TC-15
Workshop on Graph based Representations in Pattern Recognition (2001), pp. 149-159.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@165 37378900-1858-4726-a06b-e54562fdc990