Started work towards incompressible liquid phase equation of state.
New equation of state created for incompressible liquids.
This has to be told its T, P, and molar volume when it is initialised,
and then these remain fixed throughout.
(The idea of using molar volume as opposed to volume was a late one,
so some of the partial derivatives may currently be inconsistent)

In the input file use something like:
		<physicalPropertyModel type="incompressibleLiquid" >
		    <molarVolume units="cm^3/mol">162</molarVolume>
		</physicalPropertyModel>
as demonstrated in the new diesel example (examples/diesel/input.xml)

Also made tiny changes to input file:
Decreased tolerance of 1,3-hexadiene to 1% (from 10%)

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@167 37378900-1858-4726-a06b-e54562fdc990