Fixed bugs in reaction families that are their own reverse. (eg. intra_H_migration)
To generate the reverse, the labels are switched around. Suppose an H (*3)
migrates from *1 to *2. For the reverse, swap the labels on *1 and *2. However,
for groups that depend on ring size you have *4, *5, *6... in the middle. These
need to be switched also! (...*6, *5, *4).

Also removed switching of intra_OH_migration that is not its own reverse,
and added some cautionary comments to the kinetics dictionaries.

I am not sure how compatible these dictionaries (with labelled intermediate
atoms) are with the Java version of RMG

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@174 37378900-1858-4726-a06b-e54562fdc990