Fixed a bug where a homolysis reaction opened a ring but expected two products.
Unimolecular homolysis (reverse of R_recombination) will match various bonds in
the ring C1-C-C-C1 but the product will just be a straight chain biradical,
rather than two separate radicals. The template expects two products, but only
one is made, so an exception was raised. (In the java version this exception
just quietly skips to the next reaction)

It now checks (reasonably thoroughly) that this specific case is the cause of
the discrepancy, before failing quietly and trying the next reaction. If the
reason for having an unexpected number of products appears not to be because of
this, then it raises an exception as before.

Also tweaked __str__ and __repr__ for ReactionFamily

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@183 37378900-1858-4726-a06b-e54562fdc990