Finished fixing graph isomorphism bugs.
Now atoms are sorted once at the start of the graph matching, according to some
function of their connectivity, so more highly connected ones are first.
Equally connected ones will be arbitrarily ordered, but this doesn't matter,
because: we then take this sorted list, and store the index of the sort in the
atom itself. Later, when graph matching (specifically, when proposing pairs) we
sort according to this stored index. Thus, it will be in the same order every
time.

This appears to fix things!

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@189 37378900-1858-4726-a06b-e54562fdc990