When reading a library, store the 'index' (the first token on the line)
This involved changes to data.Library and reaction.ReactionFamily and
thermo.ThermoDatabase (which both inherit from data). When first reading the
library file, the data is stored in an (index, data) tuple. Both index and data
are strings at this point. The later processing into thermodynamics or kinetics
objects spits this tuple apart, processes the data, then gives the appropriate
object an index attribute. This is used in __repr__ and __str__ functions, to
help locate where data came from in a library file.

git-svn-id: https://rmg.svn.sourceforge.net/svnroot/rmg/trunk@210 37378900-1858-4726-a06b-e54562fdc990