Reaction recipes can now be applied to structures containing wildcards.
This is needed for generating product template structures automatically from
the reactant templates and the reaction recipe. Each action in the recipe is
still constrained to systems where it can physically occur; in the case of
wildcards it must be applicable to all possible values of the wildcard or an
exception is raised.