Switched to a pure Python approach to Cython (affects rmg.chem and rmg.graph).
This means that the .pyx files are returned to .py files and all Cython syntax
is removed, making the .py files valid Python. These can then be executed
without compiling the Cython if desired. The .pxd files are used to provide
type information about class attributes and methods and other variables and
functions.