Created thermo.ThermoWilhoitData class, new conversion functions, and unit tests.
There's now a ThermoWilhoitData class in thermo. Currently only getHeatCapacity
is implemented but the rest should be easy for someone (Greg?) to add. I
replaced the conversion function from GA to NASA with two functions,
convertGAtoWilhoit and convertWilhoitToNASA. These are placeholders and need to
be written so that they work. I made some more unit tests in thermotest.py,
that currently fail because the stuff's not been implemented.

Does having dozens of failing unit tests make it harder to spot genuine regressions?
Discuss... :-)