Unit test for counting rotors that the current algorithm gets wrong.
Greg thought up this test case that the current algorithm (copied from RMG Java)
does not get correct::
 start with something simple:    H3C---CH3
 now lengthen that middle bond: H3C-C#C-CH3

The first species is correctly identified as having one rotor. The second
species is the same shape, but we've lengthened the middle bond by inserting a
rigid, straight, -C#C- in the middle. It should also have one rotor, but is
identified as having two.