Got Wilhoit thermo (minimally) working.
A few bug fixes, as well as some changes to the unit tests, and most of them
pass. For propane we can match H within 2 kJ/mol throughout, but not within
1kJ/mol (at 1500K). How accurate should it in fact be? I guess the infinite-T
limit is influencing things so maybe we don't *want* to match Cp1500 spot on.

B is now defined in the ThermoWilhoitData class and not within each instance.
I and J, integration constants, have been added. We first make a ThermoWilhoitData
with 0 values for I,J, evaluate H(298) and S(298) using this, and figure
out what I and J should be to get these right.

Most amusing bug: y = t/t+B was meant to be y = t/(t+B)

Major refactoring is recommended. I think a lot of the WilhoitInt0 type functions
should be within the ThermoWilhoitData class.  I'm not sure how the NASA fitting
depends on them though, so haven't moved them yet.