Now have three separate Critical Flux Tolerances
Input file looks something like::

	<fluxTolerance>
		<keepInEdge>1e-4</keepInEdge>
		<moveToCore>0.1</moveToCore>
		<interruptSimulation>1.0</interruptSimulation>
	</fluxTolerance>

Edge species fluxes are tracked relative to the characteristic flux at that
time throughout the simulation. The highest this value reaches is stored for
later analysis.

If it exceeds fluxTolerance_interruptSimulation the simulation is interrupted
and the species is added.

If it exceeds fluxTolerance_moveToCore then it will (eventually) be added to
the core at the end of the simulation, but only one at a time (largest first)

If it falls below fluxTolerance_keepInEdge then it will be removed from the
edge, along with the reactions that involve it