Collected hardcoding of electron states into the chem module.
This was done by creating attributes for each electron state that define the
state resulting from a relevant chemical action, in this case either an
INCREMENT_RADICAL or DECREMENT_RADICAL action. (In the current RMG parlance
these are GAIN_RADICAL and LOSE_RADICAL, respectively.) None is used to
indicate a nonsensical action, e.g. decrementing the radical count past zero.

This commit breaks pickling (more) because the pickle module can't seem to
handle the new attributes, which create a nested loop of references. This
will have to be fixed before restart files will work.