Changed applyRecipe() to modify the atom types as the recipe is applied.
This is to better align with how RMG Java works. Prior to this, RMG Py would
simply recompute the atom types of the entire structure after the reaction
recipe was applied. Not only is this likely to be a bit faster, it enables
us to apply recipes to structures containing wildcards, e.g. the template
structures of the reaction family trees.

This was accomplished by introducing attributes of the AtomType class that
delineate what the atom type becomes as a result of each chemical action.
For atom types the relevant actions are FORM_BOND, BREAK_BOND,
INCREMENT_BOND, and DECREMENT_BOND. (In current RMG terminology these are
FORM_BOND, BREAK_BOND, and CHANGE_BOND.) The Atom class now has methods
for changing the atom type(s) that are called in applyRecipe().

You can enforce unique atom types on each Atom via the unique flag.