Created generateProductTemplate() to ReactionFamily.
This function can generate the product template structures automatically from
the reactant template structures and the reaction recipe. This is how RMG
Java does it; up until now RMG Py required the product structures defined
explicitly in the dictionary. The next commit will remove these structures.

The procedure for this is a bit involved due to the use of unions in several
reaction families. The product template is only allowed to have one
structure per product, and only one level of children for unions. This is
fine because we won't be using these structures to determine kinetics, only to
identify reactions.