Added reaction unit test that used to fail, and fixed it.
Problem was when unimolucular homolysis (reverse of R_recombination) broke
a bond that was in a ring, so didn't form two products. We were correctly
identifying that this was the reason for having too few products, and
returning 'None' in an attempt to signal that it should just continue
with the next reaction, but were not ready to catch this signal.

Perhaps we should use exceptions of some sort in situations like this
instead of returning 'None'.