Write Cantera CTML files for each simulation.
I had to edit the ctml_writer.py file, so have moved it from ../external
into the rmg folder as it is no longer just a copy of the Cantera release.
We should be careful about the header in it though, in terms of copyright etc.

ThermoNASA..., Species, and Reaction classes all have toCantera() methods,
which create ctml_writer instances and add them to the stored global list
of ctml_writer._species and ctml_writer._reactions.

It seems to work, although the CTML files have not actually been tested yet.