Removed product structures from kinetics dictionaries/trees.
These are no longer needed due to now having the ability to generate them
automatically when the database is loaded. The trees are now restored to as
they are in RMG Java. The dictionaries would be, except that we've removed the
CO atom type; however, they are possibly still compatible with RMG Java.