Cache some structure information (eg. connectivity values)
This reduces the number of times these properties are recalculated.
The cache is reset with a call to resetCachedStructureInfo()  (on either a
structure or graph) which loops though each atom and sets all the connectivity
values and order_label equal to -1.

In VF2_isomorphism we check if the first atom has an order_label of -1 and if
so we recalculate the connectivities and ordering.