Fix bigsmall bug in edge trimming.
Big because it had a large effect; small because it was just a couple of tabs!

It used to only consider removing a species from the edge if it had the
highest flux found so far (which obviously doesn't often hit a species
with a very low flux!).  It now discards an awful lot more edge species.