Don't abort when you find 1 product instead of 2
This occurs when you match a reaction that would break a bond, and it expects
to form two products, but they're somehow joined via another part of the
cyclic species.

For now, we just return 'None' and the reaction is ignored.

It *should* be covered by another type of (unimolecular) reaction family.