Moved restart file loading to more appropriate location.
Added species.speciesList and reaction.reactionList to items being pickled.
With this change, restart file loading is now working for the specific case
of restarting a simulation that had completed successfully. In this case,
the simulation is run again to successful completion, the model is deemed
valid, and the RMG job completes. (Note that it currently also goes through
the pruning step, so the edge size may change slightly as a result.)

Restarting interrupted jobs is not yet working. Soon!