A few more modifications that combine to make restart work.
In Atom and Bond, AtomType, ElectronState, and BondType objects are pickled
by label rather than the entire object. When unpickled, the Atom and Bond
objects are reconstructed to have the new AtomType, ElectronState, and
BondType memory locations, making our graph comparisons valid and only
keeping one copy of these in memory. Also, this shouldn't be worse in terms of
restart file size.

After unpickling species.speciesList, species.speciesCounter must be set.
Currently it is set to len(species.speciesList), although this is not
guaranteed to be safe.

In makeNewReaction(), Reaction.family objects are compared by label rather
than by memory address, as the addresses don't match. Note that this implies
that we have two copies of the kinetics database in memory (or more if
restarting multiple times). We should figure out a way around this.