Added first version of frequencies database.
This database is not backwards compatible with the Java version because of
the extensive use of {} for multiple radical numbers. The dictionary is
about 5x shorter thanks to this being allowed.

Also new is the library file. Besides allowing the frequency database to be
loaded as a standard dictionary-tree-library database, it allows us to
avoid hardcoding, which is fairly prevalent in the current RMG Java
implementation.

A tree is provided, but will likely be ignored during at least the first
implementation in Python.