Created function generateSpectralData() to do what "Frankie" does.
Currently the function is able to find the frequencies from each functional
group in the frequencies database. Fitting remaining frequencies and
hindered rotors to heat capacity data is not yet implemented.

A symmetry number-like column was added to the frequencies database library
to correct for multiple subgraph matches of equivalent groups.

A short function was added to Structure for counting the number of internal
rotor modes in a structure.