Escape thermo comments for cantera to avoid xml problems with <> characters.
We pass thermo group comments through to the cantera species objects, but some
of these comments may have < characters in, which screws up the xml of the
cantera file. The ctml_writer should escape this, but it doesn't, so we now
escape it in species.py before passing to cantera. (< becomes &lt;)