Added module spectralfit for fitting of spectral data to heat capacities.
This module is a port of "frankie", a Fortran code originally used in RMG
Java for estimating the spectral data of chemical species. Part of that
functionality was implemented in previous commits; this is what was remaining.
The original "frankie", calc_freq_code.f90, is untouched at this time, and it
depends on dqed.f90.

A new, short Fortran file spectralfit.f90 was added to act as the glue between
the Python interface and the Fortran code. This is compiled with f2py
(distributed with NumPy) and linked to the other modules. The makefile has
been updated to handle this.