Created rmg.system package for modules containing various reaction systems.
This is the first step in adding multiple reactor types to RMG. The idea is
to have Python classes that manipulate Cantera to produce the desired reactor
type. Each class needs a fromXML() function to use when parsing RMG input and
a toCantera() function for creating the reactor in Cantera.

Reactor types should derive from the base class rmg.system.base.ReactionSystem,
which will contain code to be shared across all reactor types.