Spectral data can now be generated as a part of makeNewSpecies().
There were a few minor issues in the spectral data estimation code that have
now been fixed. Most notably, I had commented out the line for one of the
cases that actually did the fitting instead of the line above it that just
printed to the console. Also changed generateSpectralData() to not assume
that thermoData is a ThermoGAData object.

There is a new option <spectralDataEstimation> in the RMG input file to turn
this on/off.