Added fromXML() methods to several classes for reading of species from XML.
This is an effort to move the RMG input file reading from one giant function
to many smaller methods of the classes that the giant function populates.
Besides cleaner code, this has the benefit of avoiding duplicate code when
RMG modules want to read input files, e.g. for pressure dependence.

The XML structure is largely based on the class/attribute hierarchy in RMG.
For example, the rmg.Species class has an attribute lennardJones for the
Lennard-Jones parameters. In XML, this is stored as

<species>
	<lennardJones>
	...
	</lennardJones>
</species>