Added functions for generating unimolecular reaction networks in makeNewReaction().
This is a port of a similar function from Java. It intelligently updates
networks so as to ensure that partial networks are properly updated as they
are filled in by future reaction generation events.

This is the first of several steps geared toward enabling pressure
dependence in the main RMG procedure.