Changed density of states to only use 10 cm^-1 grains for Beyer-Swinehart step.
The slow step in the previous implementation of the density of states
estimation was by far the convolution of multiple hindered rotor modes with
the active K-rotor. This step does not need the 10 cm^-1 grains that the
Beyer-Swinehart algorithm needs, so the fine graining was confined to just
around the Beyer-Swinehart algorithm. This results in a significant speedup,
especially when coarser grainings are desired, without much difference in
the resulting density of states.