PDepReaction objects now added to model edge after unirxn network processing.
If the network is updated, these objects should automatically update their
kinetics. However, the updated kinetics are not yet provided to Cantera for
use. Indeed, Cantera's support for multiwell networks is disappointing at this
time.

The PDepReaction objects contribute to the decision of which species to
add to the core. The network leak fluxes are not yet checked, however, so
resulting models are probably not valid.

PDepReaction objects store an attribute pointing back to the network they came
from, in case that is ever useful in the future.