Added method for updating PDepReaction kinetics in Cantera.
The method requires storing a reference to the Cantera reaction in
PDepReaction, which allows us to modify/replace the Cantera reaction when
the PDepReaction kinetics change (i.e. due to a new unirxn calculation).

Haven't been able to thoroughly test this yet, so there may be a bug or two.