Added activated species algorithm to unimolecular network handling.
This algorithm checks the leak flux of all unimolecular networks in the same
way the edge species fluxes are checked to determine model validity. This
provides a neat error bound on the partial networks that allow us to handle
networks efficiently.

These modifications have not been thoroughly tested yet, so there may be a few
bugs.