Added capability for writing of unirxn XML input files based on loaded data.
This will be useful in debugging the unirxn module, as we can write input
files for networks that fail and revisit them after execution to pinpoint the
error.

A lot of toXML() functions were added to complement the fromXML() functions
added some time ago. These will be useful as other modules are added to
RMG, as it will streamline the design/implementation of input files for these
modules.