Fixed bug in reservoir state method.
I had incorrectly assumed that portions of the steady-state vector pa that
were unused would always have a value of 0.0, when in reality I could not
depend on that being the case. The method has been adjusted to only utilize
the active-state portion of pa, which is safe and probably slightly faster
as well.