Added activated species algorithm for partial unirxn network enlargement.
Also added code to handle duplicate reactions in Cantera, which is more likely
to happen when running with pdep on (but presumably still possible even
without).

Also added workaround for networks that don't have any dissociation reactions,
as this would result in a matrix with a dimension of zero being passed from
Python to Fortran. f2py doesn't like this.