Updated RMG input file format to allow for specifying unirxn options.
The style of items in <optionList> has changed from, e.g.

<drawMolecules>on</drawMolecules>

to

<drawMolecules/>

In the new case, to turn this option off you would delete it (or comment it
out). This was done to allow for child elements when an option is turned on.

For unirxn you must specify a method, temperature and pressure lists, one or
both of grain size and/or number of grains, and an interpolation model.