Updated edge pruning to also process unirxn networks; fixed bug.
Networks that are empty (i.e. have no net or path reactions) after pruning
are removed from the list of networks. The pruned networks are *not* marked as
invalid, as the rates of the removed net reactions should be negligible and
therefore not affect the other net reactions as the rates are redistributed.

There were actually two bugs in the original edge pruning function
removeSpeciesFromEdge(). First, the function was identifying all the reactions
that should *not* be pruned, not the ones that should be. Second, the function
was only removing the last of the identified reactions. Both have been fixed.