Fixes for multiple bugs relating to unirxn systems.
The manner in which I had originally been adding unimolecular reactions to
networks and not the model edge was causing most species to not make it to
the model edge. The flow for this has been adjusted to correctly add species
to the edge and path unimolecular reactions to the appropriate network.

Occasionally the list of isomers in a network was not matching that coming
from the list of path reactions. To fix this, the isomers are now regenerated
when the network is updated.

Also fixed bug in Network.determineEnergyGrains wherein an iterator was
trampling over an important local variable.