Refactoring of unirxn: moved several things to CoreEdgeReactionModel class.
Instead of having a global list of Network objects stored in the variable
rmg.reaction.networks, we now store the list of networks in the
CoreEdgeReactionModel class under the attribute unirxnNetworks. Accordingly,
the global methods rmg.reaction.addReactionToUnimolecularNetworks() and
rmg.reaction.updateUnimolecularReactionNetworks() are now methods of the
CoreEdgeReactionModel class. This makes so much sense that I'm not entirely
sure why I didn't do it this way from the start.

As a side benefit to the above, since the reaction model and all its
attributes are being saved to the restart file, this move should enable
restart functionality when pressure dependence is on.

Other modules have been modified in response to the above changes.