Fixed restart save/load to actually work the way it should.
The key lesson is that you must pickle and unpickle all of the objects in
one concerted dump/load. If you don't, then references won't match across
objects, e.g. the Species object in the model core will point to a different
location in memory than the Species object in the global list of species.
Pickling all-at-once is also faster and uses less space on disk (and less
memory when the restart file is loaded, of course).