Fixed bug caused by trampling of Fortran modules on one another.
The key lesson here is that subroutine names must be distinct, even in
separate Fortran modules. If not, there seem to be circumstances under which
Python will attempt to call the wrong one, and a segfault will result.

This commit allows unirxn networks to be run using the reservoir state method
to successfully fall back to the modified strong collision method if the
reservoir state method fails for any reason. (However, this seems to occur
more often than I was expecting, so some investigation is required.)