Now properly processing units of A factor when loading seed mechanism.
The seed mechanism stores the preexponential in units that are the appropriate
combination of cm^3, mol, and s. RMG requires the units to be m^3, mol, and s,
and so the appropriate steps are taken to convert. Also, a new attribute
thirdBody has been added to the Reaction class to allow for third-body
reactions to be handled properly.

Also moved edge pruning to only occur when the simulation is terminated but
the model is not valid. There is no point to pruning if the model is valid
because we're done with the memory-intensive steps at that point.