Several minor bugfixes.
In unirxn networks, it is now allowed for net reactions to return kinetics
of all zeros at every temperature and pressure. This was done under the
presumption that the zeros indicate a negligible flux.

For batch reactors, added code to properly set the simulation absolute and
relative tolerances, and removed catchall exception handling around Cantera
execution code.