Significant update of documentation to reflect recent work.
New chapters were added to the developer's guide for the rmg.system and
rmg.unirxn packages. New sections were added for the rmg.settings and
rmg.spectralfit modules. A pass was made to ensure classes and functions were
up-to-date. Exceptions were moved to a new subsection and now should use the
autoexception directive.