Fixed bug in simulation step that occurs when model core has no reactions.
The Cantera integrator does not destruct cleanly if we don't let the solver
take at least one time step. This isn't a problem when there are no core
reactions because the solver will either have a large flux right away or will
quickly reach the termination time due to no transients existing in the
system. It also causes the code to be a little simpler by doing away with
the vestigial first_step construct.