Added thermo data for GRI-Mech 3.0 and capacity for loading as seed mechanism.
This was needed because RMG doesn't predict thermo data very well for a few
GRI-Mech species. In particular, the thermo data for gaseous carbon atoms was
off by hundreds of kcal/mol. Also, for GRI-Mech it is important to use all of
the parameters together as much as possible, as they are fit as a group and
may not reflect the real values.

The thermo data is added to the primary thermo library. Some work may need to
be done in the future to ensure proper loading order when multiple primary
thermo libraries are used.