Changed Arrhenius fitting to use multiple linear least-squares fits.
A naive linear least-squares can cause too much of the temperature dependence
to be placed in the A*T^n prefactor and not enough in the activation energy.
This algorithm, suggested by Franklin Goldsmith, attempts to isolate and
estimate the n parameter first before fitting the A and Ea.

The unirxn module has been updated to allow use of PDepArrhenius as an
interpolation model once again. However, the above change seems to still be
giving overflow errors, but this time due to small A and large n rather than
large A and small n.